Implement public Python API
